planetsimulator v1 without interferences and without animation

+++ Planetsimulator/Planet.py
@@ -2,7 +2,7 @@ class Planet(object):
     """Represents a planet"""
     
     #basic constructor to initialize a new planet with its attributes
-    def __init__(self, description:str, xPosition:int, yPosition:int, speed:float, mass:float):
+    def __init__(self, description:str, xPosition:float, yPosition:float, speed:float, mass:float):
         self.description = description
         self.xPosition = xPosition
         self.yPosition = yPosition
